ibox-tools
skeleton:
* close
* minimize
