+++ src/index.html
@@ -18,7 +18,9 @@
     <pwc-ibox>
       <pwc-ibox-title>
         <h5>Title</h5>
-        <pwx-ibox-tools></pwx-ibox-tools>
+        <pwc-ibox-tools minimize-button="true" close-button="true">
+          Tools
+        </pwc-ibox-tools>
       </pwc-ibox-title>
       <pwc-ibox-content>Content</pwc-ibox-content>
       <pwc-ibox-footer>Footer</pwc-ibox-footer>
